+++ src/lib/sendbird/sendbird.helper.ts
@@ -1,13 +1,13 @@
-import { HttpService } from '@nestjs/common';
+import { HttpService } from '@nestjs/axios';
 import { AxiosResponse } from 'axios';
 import * as FormData from 'form-data';
-import { Observable } from 'rxjs';
+import { lastValueFrom, Observable } from 'rxjs';
 
 export class SendBirdHelper {
   constructor(protected readonly http: HttpService) {}
 
   protected async wrapper<T>(req: Observable<AxiosResponse<T>>) {
-    return req.toPromise().then(x => x.data);
+    return lastValueFrom(req).then(x => x.data);
   }
 
   protected getFormData(data: { [key: string]: any }, fileName: string) {