+++ src/lib/socket/socket.filter.ts
@@ -1,5 +1,6 @@
 import { ArgumentsHost, Catch } from '@nestjs/common';
 import { BaseWsExceptionFilter, WsException } from '@nestjs/websockets';
+import { Socket } from 'socket.io';
 
 @Catch(WsException)
 export class ExceptionFilter extends BaseWsExceptionFilter {
@@ -7,7 +8,7 @@ export class ExceptionFilter extends BaseWsExceptionFilter {
     // const data = host.switchToWs().getData();
     // console.log('exception', exception);
     // console.log('data', data);
-    const socket = host.switchToWs().getClient<SocketIO.Socket>();
+    const socket = host.switchToWs().getClient<Socket>();
     socket.emit('exception', {
       status: 'error',
       message: exception.message || `It's a message from the exception filter`