+++ src/lib/wowza/wowza.helper.ts
@@ -1,11 +1,11 @@
-import { HttpService } from '@nestjs/common';
+import { HttpService } from '@nestjs/axios';
 import { AxiosResponse } from 'axios';
-import { Observable } from 'rxjs';
+import { lastValueFrom, Observable } from 'rxjs';
 
 export class WowzaHelper {
   constructor(protected readonly http: HttpService) {}
 
   protected async wrapper<T>(req: Observable<AxiosResponse<T>>) {
-    return req.toPromise().then(x => x.data);
+    return lastValueFrom(req).then(x => x.data);
   }
 }